nordlys/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nordlys"
version = "0.6.0"
description = "Intelligent LLM Infrastructure with Smart Model Selection"
authors = [{ name = "Botir Khaltaev", email = "botirkhaltaev@llmadaptive.uk" }]
requires-python = ">=3.11,< 3.14"
readme = "README.md"
license = "MIT"
keywords = ["llm", "ai", "model-selection", "machine-learning", "nordlys"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = [
  # Data Validation
  "pydantic>=2.11.5,<3",
  # Data Processing
  "pandas>=2.0.0",
  "polars>=1.0.0",
  # ML
  "scikit-learn>=1.5.0,<1.7.0",
  # Clustering
  "umap-learn>=0.5.0",
  "numba>=0.56.0",     # Required for Python 3.11+ compatibility
  # Caching
  "cachetools>=5.0.0,<6",
  "mpmath>=1.3.0",
  "sympy>=1.13.3",
  # Progress
  "tqdm>=4.66.0",
]


[project.urls]
Homepage = "https://www.nordlyslabs.com"
Repository = "https://github.com/Nordlys-Labs/nordlys"
Documentation = "https://docs.nordlyslabs.com"
"Bug Tracker" = "https://github.com/Nordlys-Labs/nordlys/issues"


[project.optional-dependencies]
cpu = [
  "nordlys-core>=0.3.1,<1",
  "torch>=2.2.0",
  "sentence-transformers>=5.2.0,<6",
]
cu12 = [
  "nordlys-core-cu12>=0.3.1,<1 ; sys_platform == 'linux'",
  "torch>=2.2.0",
  "sentence-transformers>=5.2.0,<6",
  "cupy-cuda12x>=13.0.0,<14.0.0",
  "cuml-cu12>=26.2.0",
]
cu13 = [
  "nordlys-core-cu13>=0.3.1,<1 ; sys_platform == 'linux'",
  "torch>=2.2.0",
  "sentence-transformers>=5.2.0,<6",
  "cupy-cuda13x>=13.0.0,<14.0.0",
  "cuml-cu13>=26.2.0",
]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu12" }, { extra = "cu13" }]]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu128", extra = "cu12", marker = "sys_platform != 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[dependency-groups]
dev = [
  "ty>=0.0.3,<1",
  "ruff>=0.11.11,<0.14",
  "types-pyyaml>=6.0.12.20250516,<7",
  "types-pyjwt>=1.7.1",
  "pytest>=9.0.2",
  "pytest-cov>=6.2.1",
  "pytest-asyncio>=0.26.0",
  "pytest-benchmark>=5.2.3",
]
test = ["pytest>=9.0.2", "pytest-cov>=6.2.1,<8", "pytest-asyncio>=0.26.0"]

[tool.uv-build]
packages = ["nordlys"]

[tool.ty]
# ty auto-discovers src/ directory and virtual environment

[tool.ty.src]
exclude = ["tests/", "scripts/"]

[tool.ty.rules]
unused-ignore-comment = "warn"
# Ignore unresolved imports globally (for C++ extension module nordlys_core_ext)
unresolved-import = "ignore"

[tool.pytest.ini_options]
addopts = "-v"
python_functions = ["test_*", "bench_*"]
filterwarnings = [
  "ignore::sklearn.exceptions.ConvergenceWarning",
  "ignore:Graph is not fully connected:UserWarning",
  "ignore:n_jobs value 1 overridden to 1 by setting random_state\\. Use no seed for parallelism\\.:UserWarning",
  "ignore::DeprecationWarning",
  "ignore:invalid escape sequence:SyntaxWarning",
  "ignore::FutureWarning",
  "ignore:.*torch.*:UserWarning",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 88
target-version = "py311"
